+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 -------------------------------------------------------------------------------------------------------------
 
-# 5.6.4 (2021-04-23)
+# 5.6.4 (2021-04-25)
 
 ### 🐣新特性
 * 【core   】     DatePattern补充DateTimeFormatter（pr#308@Gitee）
@@ -13,6 +13,7 @@
 * 【core   】     BeanUtil增加copyToList方法（issue#1526@Github）
 * 【extra  】     MailAccount增加customProperty可以用户自定义属性（pr#317@Gitee）
 * 【system 】     SystemUtil.getUserInfo()中所有平台路径统一末尾加/（issue#I3NM39@Gitee）
+* 【http   】     新增HttpDownloader，默认开启自动跳转（issue#I3NM39@Gitee）
 
 ### 🐞Bug修复
 * 【db     】     修复SQL分页时未使用别名导致的错误，同时count时取消order by子句（issue#I3IJ8X@Gitee）